Add dot-png command
With most defaults we usually use `dot` with.

Also refactor CText into own module.
